chore(web): blog overhaul
- make blog work again -> use new content layer stuff
- prepare blog series
- make sure drafts are not shown in featured list
- fix bun.lockb file by setting registry explicitly to "npmjs"
