+++ projects/rawkode.academy/platform/web/.npmrc
@@ -1,0 +1,1 @@
+registry=https://registry.npmjs.org
