+++ projects/rawkode.academy/platform/web/package.json
@@ -18,7 +18,7 @@
     "@astrojs/react": "4.0.0",
     "@astrojs/sitemap": "3.2.1",
     "@astrojs/tailwind": "5.1.3",
-    "@astrojs/ts-plugin": "^1.10.4",
+    "@astrojs/ts-plugin": "1.10.4",
     "@astrojs/vue": "5.0.1",
     "@calcom/embed-react": "1.5.1",
     "@grafana/faro-web-sdk": "1.12.2",
@@ -54,8 +54,8 @@
     "tailwindcss": "3.4.16",
     "typescript": "5.7.2",
     "utf64": "1.0.3",
-    "vue": "3.5.13",
     "vitest": "2.1.6",
+    "vue": "3.5.13",
     "vue-boring-avatars": "1.4.0",
     "vue-word-highlighter": "1.2.5",
     "vue-writer": "2.0.2"
