+++ projects/rawkode.academy/platform/web/src/components/blog/featured/article.astro
@@ -8,7 +8,7 @@ type Props = {
 const { article } = Astro.props;
 ---
 
-<a href={`/blog/${article.slug}`}>
+<a href={`/blog/${article.id}`}>
 	<article
 		class="p-4 bg-white rounded-lg border border-gray-200 shadow-md dark:bg-gray-800 dark:border-gray-700"
 	>
