+++ projects/rawkode.academy/platform/web/src/components/blog/featured/section.astro
@@ -4,6 +4,7 @@ import FeaturedArticle from "./article.astro";
 
 const latestBlogPosts = (await getCollection("blog"))
 	.slice(0, 3)
+	.filter(article => !article.data.isDraft)
 	.sort((a, b) => b.data.publishedAt.getTime() - a.data.publishedAt.getTime());
 ---
