+++ projects/rawkode.academy/platform/web/src/content/blog/how-to-ask-for-help.mdx
@@ -1,6 +1,5 @@
 ---
 title: How to ask for help
-slug: how-to-ask-for-help
 subtitle: "Maximize your chances for help: Learn to ask questions that people want to answer."
 publishedAt: 2023-10-31
 updatedAt: 2023-11-10
