+++ projects/rawkode.academy/platform/web/src/content/blog/talos-on-hetzner/01-intro.mdx
@@ -1,0 +1,10 @@
+---
+title: "Intro"
+publishedAt: 2024-07-05
+isDraft: true
+series: talos-on-hetzner
+authors:
+  - icepuma
+---
+
+Intro
