+++ projects/rawkode.academy/platform/web/src/content/blog/zed-show-and-tell.mdx
@@ -1,6 +1,5 @@
 ---
 title: "Zed: Show & Tell"
-slug: zed-show-and-tell
 subtitle: "Zed: A swift, modern code editor. Offers customization, real-time collaboration, and AI assistance."
 publishedAt: 2024-07-05
 isDraft: false
