+++ projects/rawkode.academy/platform/web/src/content/series/talos-on-hetzner.mdx
@@ -1,0 +1,7 @@
+---
+title: "Talos on Hetzner"
+authors:
+  - icepuma
+---
+
+Deploy [Talos](https://www.talos.dev/) on [Hetzner Cloud](https://www.hetzner.com/cloud) using [OpenTofu](https://opentofu.org/).
