+++ projects/rawkode.academy/platform/web/tsconfig.json
@@ -4,6 +4,11 @@
   "exclude": ["dist"],
 
   "compilerOptions": {
+    "plugins": [
+      {
+        "name": "@astrojs/ts-plugin"
+      }
+    ],
     "jsx": "react-jsx",
     "jsxImportSource": "react",
     "baseUrl": ".",
